requirements.txt
asgiref==3.8.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
decouple==0.0.7
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.14.0
gunicorn==22.0.0
idna==3.8
numpy==1.26.4
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
